Fix compilation warnings

+++ src/containers/List.h
@@ -348,7 +348,7 @@ class List {
            if (mSize != list.mSize) return false;
 
            T* items = static_cast<T*>(mBuffer);
-            for (int i=0; i < mSize; i++) {
+            for (size_t i=0; i < mSize; i++) {
                 if (items[i] != list[i]) {
                     return false;
                 }
